pyproject.toml
[project]
name = "cellenium-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "mcp",
    "crewai",
    "google",
    "google-api-core",
    "google-auth",
    "pytest",
    "python-dotenv",
    "selenium",
    "webdriver-manager",
    "requests",
    "playwright",
    "pydantic",
    "pydantic-settings",
    "gspread>=6.2.1",
    "crewai_tools",
    "crewai-tools[mcp]",
    "langchain",
    "langchain-openai>=0.2.1,<0.3.0",
    "mcp",
    "mcp-agent",
    "mcpadapt",
    "python-dotenv",
    "requests",
    "uvicorn",
    "pydantic",
    "aiohttp",
    "aiofiles",
    "logfire",
    "pytest",
]
